api/requirements.txt
pillow==12.1.0
Flask==3.1.2
markdown==3.10.2
websocket-client==1.9.0
requests==2.32.5